raspberry_pi_app/requirements.txt
# Sorting Machine - Raspberry Pi GUI Application
# Requirements

PyQt5>=5.15.0
requests>=2.28.0
python-dotenv>=1.0.0

# สำหรับ Raspberry Pi GPIO (uncomment เมื่อใช้บน Pi)
# RPi.GPIO>=0.7.0
